moved transit import methods
